Add files via upload
We cannot avoid the presence of currency in our daily lives. Anybody finds it difficult to understand and detective any currency with ease.The accuracy of currency recognition will be greatly increased, and people's burden will be effectively reduced, if computers or other systems are able to recognize counterfeit currency. In order to properly point out the currencies, our project offers an improved model along with an examination of the detection of Indian currencies. Convolutional neural networks using the Deep Learning approach have increased the accuracy and efficiency of cash recognition.In essence, the model we worked on divides money notes into different denominations.